megaparsack-indentation-shrubbery-test/corpus/pyproject.toml
[project]
name = "corpus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiolimiter>=1.2.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[tool.basedpyright]
typeCheckingMode = "recommended"
